Issue #68 added slackback to adsws
The micro-service slackback has been incorporated within the services of the
adsws.

The layout was modified to be similar to that of the accounts.

There is a set of tests accompanying the modifications.

To set this up appropriately, you have to have the correct end point for the
slack chatroom. You also need to setup a google recaptcha. There is an accompanying
feedback form to try out the posting.