Hotfix issue #81
Cannot add fix for Flask-Login > 0.2.11, because Flask-Security has not
been updated yet. Therefore, rolled back any changes based on the newer
Flask-Login.

Added tests that ensures the database is only accessed for tokens if the
user actually has an oauth client. Otherwise, this is unnecessary overhead.

Added tests for the above.