Feature / Issue #89: Forward custom headers
Added headers in the ProxyView response, such that custom headers
returned by a micro-service are retained. This is useful in certain
scenarios where we may want to return zip files, or other packets
that are not necessarily JSON.

Temporarily fixed tests that were failing in test_discoverer. This
was due to a dependency conflict. Further effort will be needed to
figure out why this really caused there to be an issue, for now the
version has been fixed.