Added solr route cookie for solr-service requests
When a request is directed to solr-service, the user token is extracted
from the request and its solr route (which link to a specific solr
instance) is retreived from redis (if it exists) before passing the
request to solr-service.

Solr-service response is also inspected for "Set-Cookie" instructions
with the solr route, in case it is present the solr route is stored in
redis using the user's token as key.